requirements_conda.txt
astropy
dill
jax
joblib
mpmath
numdifftools
numpy
pandas
pathos
python-lal
python-lalsimulation
scipy
sympy
